civit_model_loader/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9